Avoid context leakage on MainActivity::StartWalletTask
Async tasks should be static classes to avoid leaks when screens are rotated or the context is lost during background execution.

We make the private task class static and make sure we only have a weak reference to MainActivity to perform any updates to the context, only when we have a reference to it. No Context references are leaked this way.

A similar issue remains with a Thread instantiatio onCreate.

This will be fixed on the next commit